# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pelican Bay API",
    "version": "1.0.0",
    "description": "Auto-generated spec (subset) for public & admin endpoints."
  },
  "servers": [
    { "url": "https://pelicanbay.vercel.app/api" },
    { "url": "http://localhost:3000/api" }
  ],
  "paths": {
    "/admin/products": {
      "get": { "summary": "List products", "tags": ["Product"], "responses": {"200": {"description": "array of products"}}},
      "post": { "summary": "Create product", "tags": ["Product"], "requestBody": {"required": true}, "responses": {"201": {"description": "created"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer"}
    }
  },
  "security": [{ "bearerAuth": [] }]
}